anupa/requirements.txt
numpy
python-dotenv
anthropic
